Upgrade to .NET 8

+++ Howest.Prog.Cia.CurrencySwapper.Cons/Program.cs
@@ -4,7 +4,7 @@ namespace Howest.Prog.Cia.CurrencySwapper.Cons
 {
     class Program
     {
-        private const double EurToUsdRate = 1.0813771; //1 EUR is altijd 1,08 USD
+        private const decimal EurToUsdRate = 1.0813771M; //1 EUR is altijd 1,08 USD
         
         static void Main(string[] args)
         {
@@ -12,15 +12,15 @@ namespace Howest.Prog.Cia.CurrencySwapper.Cons
             Console.Write("Enter amount (EUR): ");
             string userInput = Console.ReadLine();
 
-            double amount;
-            if (!double.TryParse(userInput, out amount) || amount < 0)
+            decimal amount;
+            if (!decimal.TryParse(userInput, out amount) || amount < 0)
             {
                 Console.ForegroundColor = ConsoleColor.Red;
                 Console.WriteLine("Enter a positive amount!");
             }
             else
             {
-                double targetAmount = amount * EurToUsdRate;
+                decimal targetAmount = amount * EurToUsdRate;
 
                 Console.WriteLine($"{amount} EUR = {targetAmount:N2} USD");
             }
